API-862: Upgrade migration script
