+++ upgrades/schema/Version_4_0_20191209111232_rename_columns_in_audit_table.php
@@ -1,0 +1,44 @@
+<?php declare(strict_types=1);
+
+namespace Pim\Upgrade\Schema;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * This migration renames "count" column in "event_count" as count is a SQL keyword
+ *
+ * @author Romain Monceau <romain@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+final class Version_4_0_20191209111232_rename_columns_in_audit_table extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        $alterTableColumnCountQuery = <<<SQL
+ALTER TABLE akeneo_app_audit RENAME COLUMN count TO event_count
+SQL;
+        $alterTableDropForeignKey = <<<SQL
+ALTER TABLE akeneo_app_audit DROP FOREIGN KEY FK_AUDIT_akeneo_app_audit_code
+SQL;
+        $alterTableColumnAppCodeQuery = <<<SQL
+ALTER TABLE akeneo_app_audit CHANGE app_code app_username VARCHAR(105)
+SQL;
+        $alterTableAddForeignKey = <<<SQL
+ALTER TABLE akeneo_app_audit 
+ADD CONSTRAINT FK_AUDIT_akeneo_app_audit_username 
+FOREIGN KEY (app_username) REFERENCES oro_user (username)
+SQL;
+
+        $this->addSql($alterTableColumnCountQuery);
+        $this->addSql($alterTableDropForeignKey);
+        $this->addSql($alterTableColumnAppCodeQuery);
+        $this->addSql($alterTableAddForeignKey);
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->throwIrreversibleMigrationException();
+    }
+}
